requirements.txt
dnspython
IPy
ipython
selenium
pyvirtualdisplay
psycopg2


django==1.3
django-annoying==0.7.6
django-debug-toolbar==0.8.5
django-extensions==0.6
django-autoroot
django-model-utils==1.0.0
South==0.7.3

#django-filter==0.5.3
#django-groupcache==0.5.0
#django-localeurl==1.4
#django-recaptcha-works==0.3.4 #moze sie przydac pozniej
#django-tables2==0.7.0
django-registration


M2Crypto
django-wsgiserver
django-secure

